PartyCommand: Extends SessionCommand instead
